- fixed  winrtendtoenddelay ( audio/video) info wrongly passed to stats
[change log]
audio endtoend delay somehow was not passed.
fixed video endtoend delay value as it was passed to wrong parameter.

also, taking this chance to make sure we only pass valid endtoend delay value to stats.

[Test case]
(1)preparation, install the vs-addin
(2) from vs, Analyze -->performance and diagnostics--> webrtc diagnostic hub ->setting, for audio/video delay select winrtendtoend day..

(3) start performance analyze for PCC,
    from PCC setting, debug, enable ntp sync, and enable ETW statstics

then, perform pcc video call, make sure audio/video winrtendtoend delay graph is normal.

+++ webrtc/voice_engine/channel.h
@@ -584,7 +584,7 @@ private:
     rtc::scoped_ptr<CriticalSectionWrapper> assoc_send_channel_lock_;
     ChannelOwner associate_send_channel_ GUARDED_BY(assoc_send_channel_lock_);
 #ifdef WINRT
-    int64_t current_endtoend_delay_ms_ GUARDED_BY(ts_stats_lock_);
+    int current_endtoend_delay_ms_ GUARDED_BY(ts_stats_lock_);
 #endif
 };
